+++ AnimatedTangle/patterns/hitomezashi.js
@@ -1,7 +1,7 @@
 import { Direction } from "../../sketchlib/pga2d/Direction.js";
 import { Point } from "../../sketchlib/pga2d/Point.js";
 import { GroupPrimitive } from "../../sketchlib/primitives/GroupPrimitive.js";
-import { LineSegment } from "../../sketchlib/primitives/LinePrimitive.js";
+import { LineSegment } from "../../sketchlib/primitives/LineSegment.js";
 import { style } from "../../sketchlib/primitives/shorthand.js";
 import { Random } from "../../sketchlib/random.js";
 import { Style } from "../../sketchlib/Style.js";
