+++ DoubleSpring/DoubleSpringSystem.js
@@ -5,7 +5,7 @@ import { GeneralizedCoordinates } from "../sketchlib/VectorSpace.js";
 import { Oklch } from "../sketchlib/Oklch.js";
 import { group, style } from "../sketchlib/primitives/shorthand.js";
 import { RectPrimitive } from "../sketchlib/primitives/RectPrimitive.js";
-import { LineSegment } from "../sketchlib/primitives/LinePrimitive.js";
+import { LineSegment } from "../sketchlib/primitives/LineSegment.js";
 import { GroupPrimitive } from "../sketchlib/primitives/GroupPrimitive.js";
 import { Direction } from "../sketchlib/pga2d/Direction.js";
 import { Point } from "../sketchlib/pga2d/Point.js";
