+++ sketchlib/music/render_score.js
@@ -3,7 +3,7 @@ import { C4, C5 } from "./pitches.js";
 import { Color } from "../Color.js";
 import { Primitive } from "../primitives/Primitive.js";
 import { GroupPrimitive } from "../primitives/GroupPrimitive.js";
-import { LineSegment } from "../primitives/LinePrimitive.js";
+import { LineSegment } from "../primitives/LineSegment.js";
 import { RectPrimitive } from "../primitives/RectPrimitive.js";
 import { group, style } from "../primitives/shorthand.js";
 import { Point } from "../../sketchlib/pga2d/Point.js";
