+++ sketchlib/primitives/LineSegment.js
@@ -1,5 +1,5 @@
-import { Direction } from "../../sketchlib/pga2d/Direction.js";
-import { Point } from "../../sketchlib/pga2d/Point.js";
+import { Direction } from "../pga2d/Direction.js";
+import { Point } from "../pga2d/Point.js";
 import { PartialPrimitive } from "./Primitive.js";
 
 /**
